requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
google-generativeai==0.3.1
langchain==0.0.310
langchain-google-genai==0.0.5
pinecone-client==2.2.4
SpeechRecognition==3.10.0
PyAudio==0.2.13